+++ src/main/java/com/medspace/application/usecase/clinicEquipment/DeleteClinicEquipmentByIdUseCase.java
@@ -1,15 +1,29 @@
 package com.medspace.application.usecase.clinicEquipment;
 
 import com.medspace.application.service.ClinicEquipmentService;
+import com.medspace.application.service.ClinicService;
+import com.medspace.domain.model.ClinicEquipment;
 import jakarta.enterprise.context.ApplicationScoped;
 import jakarta.inject.Inject;
+import jakarta.transaction.Transactional;
+import jakarta.ws.rs.ForbiddenException;
 
 @ApplicationScoped
 public class DeleteClinicEquipmentByIdUseCase {
     @Inject
     ClinicEquipmentService clinicEquipmentService;
+    @Inject
+    ClinicService clinicService;
 
-    public void execute(Long id) {
+    @Transactional
+    public void execute(Long id, Long userId) {
+        ClinicEquipment clinicEquipment = clinicEquipmentService.getEquipmentById(id);
+        Long clinicId = clinicEquipment != null ? clinicEquipment.getClinic().getId() : null;
+        Boolean isOwner = clinicService.validateClinicOwnership(clinicId, userId);
+        if (!isOwner) {
+            throw new ForbiddenException("Delete unauthorized");
+        }
+
         clinicEquipmentService.deleteEquipmentById(id);
     }
 }
